refactor(kubernetes): expose image committer package
