+++ kubernetes/Dockerfile.image-committer
@@ -26,7 +26,7 @@ RUN  GOPROXY=https://goproxy.cn,direct go mod download
 
 # Copy source code
 COPY cmd/image-committer/ cmd/image-committer/
-COPY internal/imagecommitter/ internal/imagecommitter/
+COPY pkg/imagecommitter/ pkg/imagecommitter/
 
 # Build binary
 RUN CGO_ENABLED=0 GOOS=linux go build -o /usr/local/bin/image-committer ./cmd/image-committer/
