+++ kubernetes/pkg/imagecommitter/image_containerd.go
