+++ kubernetes/pkg/imagecommitter/orchestrator.go
