+++ kubernetes/pkg/imagecommitter/registry_containerd.go
