+++ kubernetes/pkg/imagecommitter/types.go
